use hoursStatusAtom to cleanup code
