+++ packages/visual-editor/src/components/contentBlocks/index.ts
@@ -4,14 +4,7 @@ export { CTAWrapper, type CTAWrapperProps } from "./CtaWrapper.tsx";
 export { Emails, type EmailsProps } from "./Emails.tsx";
 export { GetDirections, type GetDirectionsProps } from "./GetDirections.tsx";
 export { HeadingText, type HeadingTextProps } from "./HeadingText.tsx";
-export {
-  HoursStatus,
-  type HoursStatusProps,
-  type HoursStatusParams,
-  hoursDayOfWeekTemplateOverride,
-  hoursCurrentTemplateOverride,
-  hoursFutureTemplateOverride,
-} from "./HoursStatus.tsx";
+export { HoursStatus, type HoursStatusProps } from "./HoursStatus.tsx";
 export { HoursTable, type HoursTableProps } from "./HoursTable.tsx";
 export { ImageWrapper, type ImageWrapperProps } from "./Image.tsx";
 export { MapboxStaticMap, type MapboxStaticProps } from "./MapboxStaticMap.tsx";
